# api/openapi.yaml
openapi: 3.1.0
info:
  title: Kredit API
  version: 0.1.0
  description: Contract-first API for Kredit production v1.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
paths:
...
  /me/notification-preferences:
    get:
      operationId: getNotificationPreferences
      summary: Read the authenticated user's required and optional communication choices
      responses:
        '200': { description: "Notification preferences" }
    put:
      operationId: updateNotificationPreferences
      summary: Update future optional-message routing and quiet hours
      parameters:
        - $ref: '#/components/parameters/RequiredIdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [preferred_channel, fallback_channel, payment_reminders_enabled, product_updates_enabled, quiet_start_hour, quiet_end_hour, timezone, expected_version]
              properties:
                preferred_channel: { type: string, enum: [email, sms, whatsapp] }
                fallback_channel: { type: string, enum: [email, sms, whatsapp] }
                payment_reminders_enabled: { type: boolean }
                product_updates_enabled: { type: boolean }
                quiet_start_hour: { type: integer, minimum: 0, maximum: 23 }
                quiet_end_hour: { type: integer, minimum: 0, maximum: 23 }
                timezone: { type: string, const: Africa/Lagos }
                expected_version: { type: integer, minimum: 1 }
                disable_required: { type: boolean, description: Always rejected when true. }
      responses:
        '200': { description: "Preferences updated" }
        '409': { description: "Required messages cannot be disabled or the version is stale" }
  /me/recovery-codes/regenerate:
    post:
      operationId: regenerateRecoveryCodes
      summary: Revoke old recovery codes and display a new one-time set
      parameters:
        - $ref: '#/components/parameters/RequiredIdempotencyKey'
      responses:
        '200': { description: "Recovery codes shown once" }
  /account-recovery/requests:
    post:
      security: []
      operationId: requestAccountRecovery
      summary: Start enumeration-safe account recovery
      parameters:
        - $ref: '#/components/parameters/RequiredIdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identifier, channel]
              properties:
                identifier: { type: string }
                channel: { type: string, enum: [email, phone] }
      responses:
        '202': { description: "The same response is returned whether or not an account exists" }
  /account-recovery/requests/{requestID}/evidence:
    post:
      security: []
      operationId: addAccountRecoveryEvidence
      summary: Add independently verified evidence to a recovery request
      parameters:
        - { name: requestID, in: path, required: true, schema: { type: string } }
        - $ref: '#/components/parameters/RequiredIdempotencyKey'
      responses:
        '200': { description: "Evidence accepted" }
  /account-recovery/requests/{requestID}/complete:
    post:
      security: []
      operationId: completeAccountRecovery
      summary: Complete an approved recovery after cooling-off
      parameters:
        - { name: requestID, in: path, required: true, schema: { type: string } }
        - $ref: '#/components/parameters/RequiredIdempotencyKey'
      responses:
        '200': { description: "Recovery completed and existing sessions revoked" }
  /me/privacy-requests:
    get:
      operationId: listMyPrivacyRequests
      summary: List the authenticated user's privacy requests
      responses:
        '200': { description: "Privacy requests" }
    post:
      operationId: createPrivacyRequest
      summary: Submit an identity-bound privacy-rights request
      parameters:
        - $ref: '#/components/parameters/RequiredIdempotencyKey'
      responses:
        '201': { description: "Privacy request received" }
  /me/privacy-requests/{requestID}/export:
    get:
      operationId: downloadPrivacyExport
      summary: Download an unexpired privacy-filtered export belonging to the user after fresh MFA
      parameters:
        - { name: requestID, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: "Protected authoritative JSON export" }
        '404': { description: "Export missing, expired, or belongs to another user" }
        '403': { description: Fresh MFA required }
        '503': { description: Export persistence unavailable }
  /me/account-recovery/{requestID}/cancel:
    post:
      operationId: cancelAccountRecovery
      summary: Cancel an active recovery from a genuine signed-in session
      parameters:
        - { name: requestID, in: path, required: true, schema: { type: string } }
        - $ref: '#/components/parameters/RequiredIdempotencyKey'
      responses:
        '204': { description: "Recovery cancelled" }
...
  /ops/account-recovery/{requestID}/review:
    post:
      operationId: reviewAccountRecovery
      summary: Independently approve with cooling-off or reject a verified recovery
      parameters:
        - { name: requestID, in: path, required: true, schema: { type: string } }
        - $ref: '#/components/parameters/RequiredIdempotencyKey'
      responses:
        '200': { description: "Recovery review recorded" }
...
  /ops/privacy-requests/{requestID}/decide:
    post:
      operationId: decidePrivacyRequest
      summary: Record a reasoned retention-aware privacy decision
      parameters:
        - { name: requestID, in: path, required: true, schema: { type: string } }
        - $ref: '#/components/parameters/RequiredIdempotencyKey'
      responses:
        '200': { description: "Privacy decision recorded" }
  /ops/privacy-requests/{requestID}/complete:
    post:
      operationId: completePrivacyRequest
      summary: Complete approved privacy work under current owner or delegated governance
      description: Fresh MFA and CSRF are required. The current owner in solo-owner mode may complete their own approval with a recorded reason. Delegated governance requires a different reviewer. Neither mode allows completing one’s own personal request.
      parameters:
        - { name: requestID, in: path, required: true, schema: { type: string } }
        - $ref: '#/components/parameters/RequiredIdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [decision_reviewer_id, expected_version]
              properties:
                decision_reviewer_id: { type: string, format: uuid }
                expected_version: { type: integer, minimum: 1 }
                reason: { type: string, minLength: 8, maxLength: 2000, description: Required for solo-owner completion. }
      responses:
        '200': { description: "Privacy request completed" }
        '403': { description: Authorized reviewer and fresh MFA required }
        '409': { description: Review version or governance does not permit completion }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: kredit_session
  parameters:
    RequiredIdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
...